Use prefix variable to find controller's plugins
Prefix is the most reliable variable to find files or
variables for a controller, so this lets you access it
wherever it is needed without using global hardcoded
variables.

This helps to search for controller's plugins in several
locations depending on environment variables and hardcoded
variables (CONTROL_PLUGIN_PATH, CONTROL_CONFIG_PATH).

This implies also a change a LUA interpreter loading step to
correctly set the package.path variables with the environment
variables, too.

Correct the missing 'extern' in function declarations.

Depends-On: Ic448ff017e6158bec05895d63688b8968b5c6434

Change-Id: I0ad19242612559d1f4b66b6f9af9e7032d4675a8
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>