Prevent crashes with wrong afb helper library
If a binding using the controller is wrongly linked against the v4
version of the afb-libhelpers library, this can cause crashes as some
APIs have changed behavior. This commit adds a safeguard against those
cases by displaying an error message.

It also uses the correct function in the error cleanup path.

Signed-off-by: Sebastien Douheret <sebastien.douheret@iot.bzh>